+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /submission-metadata               @uk.gov.hmrc.nrsretrieval.controllers.NonrepRetrievalController.search()
+GET        /submission-metadata               @uk.gov.hmrc.nrs.retrieval.controllers.NonrepRetrievalController.search()
